Revert spike style
This reverts commit c25cad5d6bba9d208f1c9b8fa2800e1557d4ed44.

Revert "[update] change type of init_pc and break_point"

This reverts commit fc7711f24771206176f552ca3fef55d881f65f9c.

Revert "[update] change type of (to|from)host_addr"

This reverts commit 44091db0412e40cef29321356c12c0d58ed425f4.

Revert "[add] add an option simple"

This reverts commit 5bfdfd64e6324c12b973b14e1d465bc306cacbb4.

Revert "[update] change cpu.pc type (u32 -> u64)"

This reverts commit dcb0dc738ff9d3c22cd03c1add4348c98edadded.

Revert "[update] update CSRs size (u32 -> u64)"

This reverts commit de6bda1afb07f039b400fe02ef11610784712be9.

Revert "[add] modify argument of CPU::new to pass isa"

This reverts commit 0b7c7021ae9e670594688d2ef4e52d525376da61.

Revert "[add] add member mask to the Register"

This reverts commit c7e25a98ed9c7b5f704d0814f4ae2b6e75e4d5ba.

+++ src/cmdline.rs
@@ -16,9 +16,9 @@ pub struct Arguments {
     pub filename: String,
     pub exe_option: ExeOption,
     pub pkpath: Option<String>,
-    pub init_pc: Option<u64>,
+    pub init_pc: Option<u32>,
     pub isa: Option<Isa>,
-    pub break_point: Option<u64>,
+    pub break_point: Option<u32>,
     pub result_reg: Option<usize>,
     pub main_args: Option<Vec<String>>,
 }
@@ -73,7 +73,7 @@ impl Arguments {
         };
 
         let init_pc = app.value_of("pc").map(|x| {
-            u64::from_str_radix(x.trim_start_matches("0x"), 16)
+            u32::from_str_radix(x.trim_start_matches("0x"), 16)
                 .expect("invalid pc\nplease set value as hex (e.g. --pc=0x80000000)")
         });
 
@@ -84,7 +84,7 @@ impl Arguments {
         });
 
         let break_point = app.value_of("break_point").map(|x| {
-            u64::from_str_radix(x.trim_start_matches("0x"), 16)
+            u32::from_str_radix(x.trim_start_matches("0x"), 16)
                 .expect("invalid break point\nplease set value as hex (e.g. --pc=0x80000000)")
         });
